requirements.txt
aiofiles==24.1.0
aioimaplib==1.1.0
aiosmtplib==3.0.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.35.3
packaging==24.1
pluggy==1.5.0
priority==2.0.0
pydantic==2.7.4
pydantic_core==2.18.4
pytest==8.2.2
python-dotenv==1.0.1
Quart==0.19.6
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
Werkzeug==3.0.3
wsproto==1.2.0
